+++ workbench/bootstrap/providers.php
@@ -2,4 +2,5 @@
 
 return [
     \Workbench\App\Providers\Filament\AdminPanelProvider::class,
+    \Workbench\App\Providers\Filament\UserPanelProvider::class,
 ];